+++ src/colorspace/xyz.rs
@@ -1,16 +1,16 @@
 #![cfg(feature = "experimental")]
 
-use super::{Color, ColorError};
+use super::{Color, ColorError, Float};
 
 #[derive(Clone, Copy, PartialEq, Debug)]
 pub struct XyzColor {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: Float,
+    pub y: Float,
+    pub z: Float,
 }
 
 impl XyzColor {
-    pub fn new(x: f64, y: f64, z: f64) -> Self {
+    pub fn new(x: Float, y: Float, z: Float) -> Self {
         Self { x, y, z }
     }
 }
